added unit tests for annotation pkg
